requirements.txt
Django==3.2.5
Jinja2==3.0.1
pbr==5.6.0
polyline==1.4.0
networkx==2.8.6
osmnx==1.2.2
djangorestframework==3.14.0
numpy~=1.23.5
requests~=2.28.1
folium~=0.13.0
scikit-learn==1.1.2
django-rest-framework==0.1.0
